requirements.txt
django
djangorestframework
djangorestframework-simplejwt
pillow
django-cors-headers
drf-spectacular
drf-spectacular[sidecar]
